fix: data format for tables and updated table component
